refactor(editor): consolidate canvas hit ordering
Place instance, route, endpoint, and annotation hit targets behind one canvas-layer boundary. These elements share SVG hit-order ownership and were always composed together by EditorCanvasSurface; co-locating them makes that ordering explicit while preserving the surface as the overall layer-order owner. Keep wiring and transient overlays separate because they represent distinct editing and preview contracts.\n\nValidation: pnpm typecheck; pnpm test:local apps/editor/src/canvas/editor-canvas-hit-layer.test.tsx apps/editor/src/canvas/editor-canvas-surface.test.tsx apps/editor/src/app/App.test.tsx; pnpm test:e2e:local apps/editor/e2e/manual-editor.spec.ts --grep deletes-a-wire/deletes-a-selected-junction/directional-marquee/drag-value-annotation; git diff --check.\n\nTest-Impact: tests-updated
